First git release
Some of us marketing and usability focused folks know through experience
and experimentation that for the general population, tabs don't work.
People are much more likely to scroll than to click.
viewtopic.php?f=10&t=22111

It is not a coincidence that Amazon use mile-long pages instead of the
neat organization of tabs.

Unfortunately, the default theme uses tabs for reviews, related
products, etc., which are critical informations that assist in
purchase-decisions. Hiding information is not a good idea.

Hence, this very simple vQmod, which essentially explodes the tabs and
shows a longer, scrollable page.

In the interest of cleanliness, tiny tweaks are made to page CSS as
well.

Because this is a vQmod, no core files are touched, and the changes
should carry over to the next version.